+++ src/__tests__/Home.test.js
@@ -1,8 +1,15 @@
 import React from 'react';
-import { render, cleanup, wait, waitFor } from '@testing-library/react';
+import { render, cleanup, waitFor } from '@testing-library/react';
 import { HomePage } from '../components/pages/Home';
 import { LoadingComponent } from '../components/common';
 import { BrowserRouter as Router } from 'react-router-dom';
+
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+import { mainReducer } from '../reducers/mainReducer';
+
+const store = createStore(mainReducer, applyMiddleware(thunk));
 
 afterEach(cleanup);
 
@@ -22,20 +29,23 @@ jest.mock('@okta/okta-react', () => ({
 describe('<HomeContainer /> testing suite', () => {
   test('mounts a page', async () => {
     const { findByText, getByText, queryByText } = render(
-      <Router>
-        <HomePage
-          LoadingComponent={() => (
-            <LoadingComponent message="...fetching profile" />
-          )}
-        />
-      </Router>
+      <Provider store={store}>
+        <Router>
+          <HomePage
+            LoadingComponent={() => (
+              <LoadingComponent message="...fetching profile" />
+            )}
+          />
+        </Router>
+      </Provider>
     );
     let loader = getByText(/...fetching profile/i);
     expect(loader).toBeInTheDocument();
 
     await waitFor(async () => {
-      await findByText(/hi sara/i);
+      await findByText(/My Budget/i);
     });
+
     loader = queryByText(/...fetching profile/i);
     expect(loader).toBeNull();
   });
