+++ src/__tests__/RenderHomePage.test.js
@@ -2,7 +2,13 @@ import RenderHomePage from '../components/pages/Home/RenderHomePage';
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { render } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+import { mainReducer } from '../reducers/mainReducer';
+
+const store = createStore(mainReducer, applyMiddleware(thunk));
 
 describe('<RenderHomePage /> test suite', () => {
   test('it handles a loading state', () => {
@@ -10,18 +16,13 @@ describe('<RenderHomePage /> test suite', () => {
       logout: jest.fn(),
     };
     const { getByText } = render(
-      <Router>
-        <RenderHomePage userInfo={{ name: 'Sara' }} authService={authService} />
-      </Router>
+      <Provider store={store}>
+        <Router>
+          <RenderHomePage authService={authService} />
+        </Router>
+      </Provider>
     );
-    const dropdown = getByText(/profile/i);
-    userEvent.click(dropdown);
-    const logout = getByText(/logout/i);
-    userEvent.click(logout);
-    expect(authService.logout).toHaveBeenCalledTimes(1);
-    expect(getByText(/logout/i).innerHTML).toBe('Logout');
-    expect(getByText(/hi sara, welcome to/i).innerHTML).toBe(
-      'Hi Sara, Welcome to SaverLife'
-    );
+    const navbar = getByText(/home/i);
+    expect(navbar.innerHTML).toBe('Home');
   });
 });
